Splitter dto og domeneobjekter for avsnitt og brev
