+++ src/main/kotlin/no/nav/familie/klage/brev/AvsnittRepository.kt
@@ -1,6 +1,6 @@
 package no.nav.familie.klage.brev
 
-import no.nav.familie.klage.brev.dto.Avsnitt
+import no.nav.familie.klage.brev.domain.Avsnitt
 import no.nav.familie.klage.repository.InsertUpdateRepository
 import no.nav.familie.klage.repository.RepositoryInterface
 import org.springframework.data.jdbc.repository.query.Modifying
