+++ src/main/kotlin/no/nav/familie/klage/brev/dto/BrevMedAvsnittDto.kt
@@ -1,0 +1,9 @@
+package no.nav.familie.klage.brev.dto
+
+import java.util.UUID
+
+data class BrevMedAvsnittDto(
+    val behandlingId: UUID,
+    val overskrift: String,
+    val avsnitt: List<AvsnittDto>?
+)
