+++ src/main/kotlin/no/nav/familie/klage/brev/dto/VedtaksbrevFritekstDto.kt
@@ -1,10 +1,9 @@
-package no.nav.familie.ef.sak.brev.dto
+package no.nav.familie.klage.brev.dto
 
-import no.nav.familie.klage.brev.dto.Avsnitt
 import java.util.UUID
 
 data class VedtaksbrevFritekstDto(
     val overskrift: String,
-    val avsnitt: List<Avsnitt>,
+    val avsnitt: List<AvsnittDto>,
     val behandlingId: UUID
 )
