+++ src/test/kotlin/no/nav/familie/klage/infrastruktur/config/OppslagSpringRunnerTest.kt
@@ -6,8 +6,8 @@ import com.github.tomakehurst.wiremock.WireMockServer
 import no.nav.familie.klage.ApplicationLocal
 import no.nav.familie.klage.behandling.domain.Behandling
 import no.nav.familie.klage.behandlingshistorikk.domain.Behandlingshistorikk
+import no.nav.familie.klage.brev.domain.Avsnitt
 import no.nav.familie.klage.brev.domain.Brev
-import no.nav.familie.klage.brev.dto.Avsnitt
 import no.nav.familie.klage.fagsak.domain.FagsakDomain
 import no.nav.familie.klage.fagsak.domain.FagsakPerson
 import no.nav.familie.klage.fagsak.domain.PersonIdent
